LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource_TAO_space-expected.txt

Description

This test validates that for a cross origin resource, the timing allow check algorithm will fail if the Timing-Allow-Origin header value list contains [origin|*]{2,}, separated by space.


PASS window.performance is defined
PASS window.performance.getEntriesByType is defined
PASS redirectStart, redirectEnd, domainLookupStart, domainLookupEnd, connectStart, connectEnd, secureConnectionStart, requestStart, and responseStart -- should be all returned as 0 when the Timing-Allow-Origin header value of the HTTP response is a space separated origin/wildcard list
